Refactor autogrouping script functions into utils file
